+++ src/components/write/write-form.jsx
@@ -3,10 +3,10 @@ import { useState, useRef } from 'react'
 import styled from 'styled-components'
 
 import Button from '../ui/button'
-import ImageIcon from '../assets/icons/image-icon.svg'
-import FolderIcon from '../assets/icons/folder-icon.svg'
-import ClipIcon from '../assets/icons/clip-icon.svg'
-import ContainerIcon from '../assets/icons/container-icon.svg'
+import ImageIcon from '../../assets/icons/image-icon.svg'
+import FolderIcon from '../../assets/icons/folder-icon.svg'
+import ClipIcon from '../../assets/icons/clip-icon.svg'
+import ContainerIcon from '../../assets/icons/container-icon.svg'
 
 const FormContainer = styled.div`
     width: 100%;
